requirements.txt
Django~=3
ex_oracle
djangorestframework 
markdown
django-filter 
django-cors-headers